+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add option `bib-label-sep` to set the space between label and entry content ([#109](https://github.com/zepinglee/citeproc-lua/issues/109)).
 
+### Changed
+
+- The following style IDs are renamed (upstream changes in CSL styles repository).
+  - `chicago-fullnote-bibliography` => `chicago-notes-bibliography`
+  - `chicago-note-bibliography` => `chicago-shortened-notes-bibliography`
+  - `modern-humanities-research-association` => `modern-humanities-research-association-notes`
+
 ## [0.8.3] - 2025-09-08
 
 ### Fixed
@@ -333,7 +340,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial CTAN release.
 
-[Unreleased]: https://github.com/zepinglee/citeproc-lua/compare/v0.8.3...HEAD
+[Unreleased]: https://github.com/zepinglee/citeproc-lua/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/zepinglee/citeproc-lua/compare/v0.8.3...v0.9.0
 [0.8.3]: https://github.com/zepinglee/citeproc-lua/compare/v0.8.2...v0.8.3
 [0.8.2]: https://github.com/zepinglee/citeproc-lua/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/zepinglee/citeproc-lua/compare/v0.8.0...v0.8.1